requirements.txt
newspaper3k
lxml_html_clean
requests
ddgs
typing_extensions
moviepy
pillow